Project card: on click open side modal
