+++ front/app/containers/Admin/inspirationHub/ProjectDrawer/Header.tsx
@@ -13,14 +13,10 @@ import { ProjectLibraryProjectData } from 'api/project_library_projects/types';
 
 import { parseBackendDateString } from 'utils/dateUtils';
 
-import {
-  useLocalizeProjectLibrary,
-  getTenantURL,
-  getProjectURL,
-  formatDate,
-} from '../utils';
+import { useLocalizeProjectLibrary } from '../utils';
 
 import ExternalLink from './ExternalLink';
+import { getTenantURL, getProjectURL, formatDate } from './utils';
 
 interface Props {
   attributes: ProjectLibraryProjectData['attributes'];
